+++ resources/lang/cs/dielo.php
@@ -28,6 +28,7 @@ return array(
     'item_attr_integrity_work' => 'integrita s díly',
     'item_attr_inscription'    => 'značení',
     'item_attr_gallery'        => 'galerie',
+    'item_attr_current_location' => 'aktuální lokace',
     'item_attr_contributor'    => 'kurátor',
     'item_attr_identifier'     => 'inventární číslo',
     'item_attr_licence'        => 'licence',