+++ resources/lang/en/dielo.php
@@ -28,6 +28,7 @@ return array(
     'item_attr_integrity_work' => 'in a series with',
     'item_attr_inscription'    => 'inscription',
     'item_attr_gallery'        => 'institution',
+    'item_attr_current_location' => 'current location',
     'item_attr_contributor'    => 'curator',
     'item_attr_identifier'     => 'inventory number',
     'item_attr_licence'        => 'licence',