+++ resources/lang/sk/dielo.php
@@ -28,6 +28,7 @@ return array(
     'item_attr_integrity_work' => 'integrita s dielami',
     'item_attr_inscription'    => 'značenie',
     'item_attr_gallery'        => 'inštitúcia',
+    'item_attr_current_location' => 'aktuálna lokácia',
     'item_attr_contributor'    => 'kurátor',
     'item_attr_identifier'     => 'inventárne číslo',
     'item_attr_licence'        => 'licencia',